+++ lib/ui/transfer/transfer_overview_sheet.dart
@@ -1,30 +1,31 @@
 import 'dart:async';
+
 import 'package:auto_size_text/auto_size_text.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:barcode_scan2/barcode_scan2.dart';
 import 'package:event_taxi/event_taxi.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_nano_ffi/flutter_nano_ffi.dart';
-import 'package:nautilus_wallet_flutter/localization.dart';
-import 'package:nautilus_wallet_flutter/dimens.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:logger/logger.dart';
 import 'package:nautilus_wallet_flutter/appstate_container.dart';
 import 'package:nautilus_wallet_flutter/bus/events.dart';
+import 'package:nautilus_wallet_flutter/dimens.dart';
+import 'package:nautilus_wallet_flutter/localization.dart';
 import 'package:nautilus_wallet_flutter/model/wallet.dart';
 import 'package:nautilus_wallet_flutter/network/account_service.dart';
 import 'package:nautilus_wallet_flutter/network/model/response/account_balance_item.dart';
 import 'package:nautilus_wallet_flutter/network/model/response/accounts_balances_response.dart';
 import 'package:nautilus_wallet_flutter/service_locator.dart';
+import 'package:nautilus_wallet_flutter/styles.dart';
 import 'package:nautilus_wallet_flutter/ui/transfer/transfer_confirm_sheet.dart';
 import 'package:nautilus_wallet_flutter/ui/transfer/transfer_manual_entry_sheet.dart';
+import 'package:nautilus_wallet_flutter/ui/util/ui_util.dart';
 import 'package:nautilus_wallet_flutter/ui/widgets/animations.dart';
+import 'package:nautilus_wallet_flutter/ui/widgets/buttons.dart';
 import 'package:nautilus_wallet_flutter/ui/widgets/sheet_util.dart';
 import 'package:nautilus_wallet_flutter/ui/widgets/sheets.dart';
-import 'package:nautilus_wallet_flutter/ui/widgets/buttons.dart';
-import 'package:nautilus_wallet_flutter/ui/util/ui_util.dart';
-import 'package:nautilus_wallet_flutter/styles.dart';
 import 'package:nautilus_wallet_flutter/util/caseconverter.dart';
 import 'package:nautilus_wallet_flutter/util/nanoutil.dart';
-import 'package:logger/logger.dart';
 
 class AppTransferOverviewSheet {
   static const int NUM_SWEEP = 15; // Number of accounts to sweep from a seed
@@ -32,7 +33,7 @@ class AppTransferOverviewSheet {
   NanoUtil? _nanoUtil;
 
   // accounts to private keys/account balances
-  Map<String, AccountBalanceItem> privKeyBalanceMap = Map();
+  Map<String, AccountBalanceItem> privKeyBalanceMap = {};
 
   bool _animationOpen = false;
 
@@ -70,7 +71,7 @@ class AppTransferOverviewSheet {
                   minimum: EdgeInsets.only(
                     bottom: MediaQuery.of(context).size.height * 0.035,
                   ),
-                  child: Container(
+                  child: SizedBox(
                     width: double.infinity,
                     child: Column(
                       mainAxisSize: MainAxisSize.max,
@@ -81,7 +82,7 @@ class AppTransferOverviewSheet {
                           crossAxisAlignment: CrossAxisAlignment.start,
                           children: <Widget>[
                             // Emtpy SizedBox
-                            SizedBox(
+                            const SizedBox(
                               height: 60,
                               width: 60,
                             ),
@@ -89,7 +90,7 @@ class AppTransferOverviewSheet {
                               children: <Widget>[
                                 // Sheet handle
                                 Container(
-                                  margin: EdgeInsets.only(top: 10),
+                                  margin: const EdgeInsets.only(top: 10),
                                   height: 5,
                                   width: MediaQuery.of(context).size.width * 0.15,
                                   decoration: BoxDecoration(
@@ -99,7 +100,7 @@ class AppTransferOverviewSheet {
                                 ),
                                 // The header
                                 Container(
-                                  margin: EdgeInsets.only(top: 15.0),
+                                  margin: const EdgeInsets.only(top: 15.0),
                                   constraints: BoxConstraints(maxWidth: MediaQuery.of(context).size.width - 140),
                                   child: AutoSizeText(
                                     CaseChange.toUpperCase(AppLocalization.of(context)!.transferHeader, context),
@@ -112,7 +113,7 @@ class AppTransferOverviewSheet {
                               ],
                             ),
                             // Emtpy SizedBox
-                            SizedBox(
+                            const SizedBox(
                               height: 60,
                               width: 60,
                             ),
@@ -128,21 +129,21 @@ class AppTransferOverviewSheet {
                               Container(
                                 constraints:
                                     BoxConstraints(maxHeight: MediaQuery.of(context).size.height * 0.2, maxWidth: MediaQuery.of(context).size.width * 0.6),
-                                child: Stack(
-                                  children: <Widget>[
-                                    Center(
-                                      child: SvgPicture.asset('legacy_assets/transferfunds_illustration_start_paperwalletonly.svg',
-                                          color: StateContainer.of(context).curTheme.text45, width: MediaQuery.of(context).size.width),
-                                    ),
-                                    Center(
-                                      child: SvgPicture.asset('legacy_assets/transferfunds_illustration_start_nautiluswalletonly.svg',
-                                          color: StateContainer.of(context).curTheme.primary, width: MediaQuery.of(context).size.width),
-                                    ),
-                                  ],
-                                ),
+                                // child: Stack(
+                                //   children: <Widget>[
+                                //     Center(
+                                //       child: SvgPicture.asset('legacy_assets/transferfunds_illustration_start_paperwalletonly.svg',
+                                //           color: StateContainer.of(context).curTheme.text45, width: MediaQuery.of(context).size.width),
+                                //     ),
+                                //     Center(
+                                //       child: SvgPicture.asset('legacy_assets/transferfunds_illustration_start_nautiluswalletonly.svg',
+                                //           color: StateContainer.of(context).curTheme.primary, width: MediaQuery.of(context).size.width),
+                                //     ),
+                                //   ],
+                                // ),
                               ),
                               Container(
-                                alignment: AlignmentDirectional(-1, 0),
+                                alignment: AlignmentDirectional.centerStart,
                                 margin: EdgeInsets.symmetric(horizontal: smallScreen(context) ? 35 : 50, vertical: 20),
                                 child: AutoSizeText(
                                   AppLocalization.of(context)!.transferIntro.replaceAll("%1", AppLocalization.of(context)!.scanQrCode),
@@ -165,8 +166,8 @@ class AppTransferOverviewSheet {
                               Dimens.BUTTON_TOP_DIMENS,
                               onPressed: () {
                                 UIUtil.cancelLockEvent();
-                                BarcodeScanner.scan(/*StateContainer.of(context).curTheme.qrScanTheme*/).then((val) {
-                                  var value = val.rawContent;
+                                BarcodeScanner.scan(/*StateContainer.of(context).curTheme.qrScanTheme*/).then((ScanResult val) {
+                                  final String value = val.rawContent;
 
                                   if (!NanoSeeds.isValidSeed(value)) {
                                     UIUtil.showSnackbar(AppLocalization.of(context)!.qrInvalidSeed, context);
@@ -208,19 +209,19 @@ class AppTransferOverviewSheet {
   Future<void> startTransfer(BuildContext context, String seed, {bool manualEntry = false}) async {
     // Show loading overlay
     _animationOpen = true;
-    AnimationType animationType = manualEntry ? AnimationType.TRANSFER_SEARCHING_MANUAL : AnimationType.TRANSFER_SEARCHING_QR;
+    final AnimationType animationType = manualEntry ? AnimationType.TRANSFER_SEARCHING_MANUAL : AnimationType.TRANSFER_SEARCHING_QR;
     AppAnimation.animationLauncher(context, animationType, onPoppedCallback: () => _animationOpen = false);
     // Get accounts from seed
-    List<String> accounts = await getAccountsFromSeed(context, seed);
+    final List<String> accounts = await getAccountsFromSeed(context, seed);
     try {
-      AccountsBalancesResponse resp = await sl.get<AccountService>().requestAccountsBalances(accounts);
+      final AccountsBalancesResponse resp = await sl.get<AccountService>().requestAccountsBalances(accounts);
       if (_animationOpen) {
         Navigator.of(context).pop();
       }
-      List<String> accountsToRemove = [];
+      final List<String> accountsToRemove = [];
       resp.balances!.forEach((String account, AccountBalanceItem balItem) {
-        BigInt balance = BigInt.parse(balItem.balance!);
-        BigInt receivable = BigInt.parse(balItem.receivable!);
+        final BigInt balance = BigInt.parse(balItem.balance!);
+        final BigInt receivable = BigInt.parse(balItem.receivable!);
         if (balance + receivable == BigInt.zero) {
           accountsToRemove.add(account);
         } else {
@@ -232,7 +233,7 @@ class AppTransferOverviewSheet {
       accountsToRemove.forEach((String account) {
         privKeyBalanceMap.remove(account);
       });
-      if (privKeyBalanceMap.length == 0) {
+      if (privKeyBalanceMap.isEmpty) {
         UIUtil.showSnackbar(AppLocalization.of(context)!.transferNoFunds, context);
         return;
       }
@@ -250,17 +251,17 @@ class AppTransferOverviewSheet {
 
   Future<BigInt> getGiftCardBalance(BuildContext context, String seed) async {
     // Get accounts from seed
-    List<String> accounts = await getAccountsFromSeed(context, seed);
+    final List<String> accounts = await getAccountsFromSeed(context, seed);
     try {
-      AccountsBalancesResponse resp = await sl.get<AccountService>().requestAccountsBalances(accounts);
+      final AccountsBalancesResponse resp = await sl.get<AccountService>().requestAccountsBalances(accounts);
       if (_animationOpen) {
         Navigator.of(context).pop();
       }
-      List<String> accountsToRemove = [];
+      final List<String> accountsToRemove = [];
       BigInt totalBalance = BigInt.zero;
       resp.balances!.forEach((String account, AccountBalanceItem balItem) {
-        BigInt balance = BigInt.parse(balItem.balance!);
-        BigInt receivable = BigInt.parse(balItem.receivable!);
+        final BigInt balance = BigInt.parse(balItem.balance!);
+        final BigInt receivable = BigInt.parse(balItem.receivable!);
         if (balance + receivable == BigInt.zero) {
           accountsToRemove.add(account);
         } else {
@@ -273,7 +274,7 @@ class AppTransferOverviewSheet {
       accountsToRemove.forEach((String account) {
         privKeyBalanceMap.remove(account);
       });
-      if (privKeyBalanceMap.length == 0) {
+      if (privKeyBalanceMap.isEmpty) {
         // return null;
         return BigInt.zero;
       }
@@ -292,18 +293,18 @@ class AppTransferOverviewSheet {
     AppAnimation.animationLauncher(context, AnimationType.TRANSFER_SEARCHING_MANUAL, onPoppedCallback: () => _animationOpen = false);
 
     // sleep for a couple seconds to flex the animation:
-    await Future.delayed(Duration(seconds: 3));
+    await Future.delayed(const Duration(seconds: 3));
     // Get accounts from seed
-    List<String> accounts = await getAccountsFromSeed(context, seed);
+    final List<String> accounts = await getAccountsFromSeed(context, seed);
     try {
-      AccountsBalancesResponse resp = await sl.get<AccountService>().requestAccountsBalances(accounts);
+      final AccountsBalancesResponse resp = await sl.get<AccountService>().requestAccountsBalances(accounts);
       if (_animationOpen) {
         Navigator.of(context).pop();
       }
-      List<String> accountsToRemove = [];
+      final List<String> accountsToRemove = [];
       resp.balances!.forEach((String account, AccountBalanceItem balItem) {
-        BigInt balance = BigInt.parse(balItem.balance!);
-        BigInt receivable = BigInt.parse(balItem.receivable!);
+        final BigInt balance = BigInt.parse(balItem.balance!);
+        final BigInt receivable = BigInt.parse(balItem.receivable!);
         if (balance + receivable == BigInt.zero) {
           accountsToRemove.add(account);
         } else {
@@ -315,16 +316,12 @@ class AppTransferOverviewSheet {
       accountsToRemove.forEach((String account) {
         privKeyBalanceMap.remove(account);
       });
-      if (privKeyBalanceMap.length == 0) {
+      if (privKeyBalanceMap.isEmpty) {
         UIUtil.showSnackbar(AppLocalization.of(context)!.transferNoFunds, context);
         return;
       }
 
       AppTransferConfirmSheet().createState().autoProcessWallets(privKeyBalanceMap, wallet);
-      //.autoProcessWallets(privKeyBalanceMap);
-      // Go to confirmation screen
-      // EventTaxiImpl.singleton().fire(TransferConfirmEvent(balMap: privKeyBalanceMap));
-      // Navigator.of(context).pop();
     } catch (e) {
       sl.get<Logger>().e("error", e);
       if (_animationOpen) {
@@ -340,18 +337,18 @@ class AppTransferOverviewSheet {
     AppAnimation.animationLauncher(context, AnimationType.TRANSFER_SEARCHING_MANUAL, onPoppedCallback: () => _animationOpen = false);
 
     // sleep for a couple seconds to flex the animation:
-    await Future.delayed(Duration(seconds: 3));
+    await Future.delayed(const Duration(seconds: 3));
     // Get accounts from seed
-    List<String> accounts = await getAccountsFromSeed(context, seed);
+    final List<String> accounts = await getAccountsFromSeed(context, seed);
     try {
-      AccountsBalancesResponse resp = await sl.get<AccountService>().requestAccountsBalances(accounts);
+      final AccountsBalancesResponse resp = await sl.get<AccountService>().requestAccountsBalances(accounts);
       if (_animationOpen) {
         Navigator.of(context).pop();
       }
-      List<String> accountsToRemove = [];
+      final List<String> accountsToRemove = [];
       resp.balances!.forEach((String account, AccountBalanceItem balItem) {
-        BigInt balance = BigInt.parse(balItem.balance!);
-        BigInt receivable = BigInt.parse(balItem.receivable!);
+        final BigInt balance = BigInt.parse(balItem.balance!);
+        final BigInt receivable = BigInt.parse(balItem.receivable!);
         if (balance + receivable == BigInt.zero) {
           accountsToRemove.add(account);
         } else {
@@ -363,7 +360,7 @@ class AppTransferOverviewSheet {
       accountsToRemove.forEach((String account) {
         privKeyBalanceMap.remove(account);
       });
-      if (privKeyBalanceMap.length == 0) {
+      if (privKeyBalanceMap.isEmpty) {
         UIUtil.showSnackbar(AppLocalization.of(context)!.transferNoFunds, context);
         return;
       }
@@ -384,7 +381,7 @@ class AppTransferOverviewSheet {
 
   /// Get NUM_SWEEP accounts from seed to request balances for
   Future<List<String>> getAccountsFromSeed(BuildContext context, String seed) async {
-    List<String> accountsToRequest = [];
+    final List<String> accountsToRequest = [];
     String privKey;
     String address;
     // Get NUM_SWEEP private keys + accounts from seed